pyproject.toml
[project]
name = "dbx-token-exchange-sample"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "respx>=0.21",
    "pylint>=3.0",
    "pytest-cov>=5.0",
    "msal>=1.35.1",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/app"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.pylint.main]
source-roots = ["src"]

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
]

[tool.uv]
package = true
